moving test to runtime moving base classes to common
