+++ Assets/Mirror/Runtime/AssemblyInfo.cs
@@ -1,5 +1,6 @@
 using System.Runtime.CompilerServices;
 
+[assembly: InternalsVisibleTo("Mirror.Tests.Common")]
 [assembly: InternalsVisibleTo("Mirror.Tests")]
 [assembly: InternalsVisibleTo("Mirror.Tests.Runtime")]
 [assembly: InternalsVisibleTo("Mirror.Tests.Performance.Editor")]
